fix(bots): implement correct agar.io protocol encryption
- Replace static 0x7999 key (UInt64) with proper version integer (UInt32)
  derived from client version string via versionStringToInt()
- Add MurmurHash2 for client→server encryption key derivation from
  server hostname + version string (seed=255)
- Add XOR encryption with rotating key for all client→server packets
  after F1 handshake (rotateKey after each send)
- Add server→client XOR decryption using movementKey ^ versionInt
- Add LZ4 decompression for server 0xFF compressed packets
- Fix world update parser: Int32 coords (was Int16), UInt16 remove count
- Fix move packet: Int32 coords + movementKey field (13 bytes, was 21)
- Remove null terminator from spawn packet to match live protocol
- Protocol: v22 handshake, version int from "26.6.0" = 260600

+++ XRDMod/PlayerInfo.swift
@@ -26,8 +26,8 @@ struct PlayerInfo: Identifiable, Hashable {
 
 struct CellUpdate {
     var id: UInt32
-    var x: Int16
-    var y: Int16
+    var x: Int32
+    var y: Int32
     var size: Int16
     var color: UInt32
     var flags: UInt8
